+++ lib/snippy/otp_check.ex
@@ -1,9 +1,11 @@
 defmodule Snippy.OTPCheck do
   @moduledoc false
 
+  alias Snippy.OtpInfo
+
   @min_otp 25
 
-  @otp_release :erlang.system_info(:otp_release) |> List.to_integer()
+  @otp_release OtpInfo.release()
 
   if @otp_release < @min_otp do
     raise CompileError,
@@ -11,7 +13,7 @@ defmodule Snippy.OTPCheck do
   end
 
   def check! do
-    otp = :erlang.system_info(:otp_release) |> List.to_integer()
+    otp = OtpInfo.release()
 
     if otp < @min_otp do
       raise "Snippy requires OTP >= #{@min_otp}, got #{otp}"